+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+### Changed
+- IO methods in ioutils now take cancellation tokens
 ## [v1.5.0.0]
 ### Added
 - #1214 Additional log entries